src/main/java/tdl/record/sourcecode/snapshot/helpers/GitHelper.java
package tdl.record.sourcecode.snapshot.helpers;

import java.io.File;
import java.nio.file.Path;
import org.eclipse.jgit.lib.RepositoryCache;
import org.eclipse.jgit.storage.file.FileRepositoryBuilder;
import org.eclipse.jgit.util.FS;

public class GitHelper {

    public static boolean isGitDirectory(Path path) {
        File directory = path.toFile();
        FileRepositoryBuilder builder = new FileRepositoryBuilder();
        builder.addCeilingDirectory(directory);
        builder.findGitDir(directory);
        return builder.getGitDir() != null;
    }
}
